pyproject.toml
[tool.poetry]
name = "task"
version = "0.1.0"
description = ""
authors = ["chanhyoYun <chloeee2@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-dotenv = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
